Extract advisor brief contract modules
